requirements.txt
fastapi~=0.115.6
uvicorn~=0.31.0
motor~=3.6.0
pydantic[email]~=2.9.2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
motor
pydantic_settings
anyio
pytest-asyncio
pytest-tornasync
pytest-trio
pytest-twisted
twisted
asgi_lifespan
pydantic[email]
pydantic-settings~=2.5.2
python-dotenv~=0.21.0
pymongo~=4.9.2
mongomock
pytest~=8.3.3
httpx~=0.27.0
asgi-lifespan~=2.1.0
bcrypt
google-cloud-storage
jinja2
aiohttp
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
